Complete the Rock Paper Scissors UI for TOP
In this commit, I added lots of functionality to the basic
rock paper scissors game from earlier in TOP curriculum.
Specifically, I added a very basic UI that allows players to
interact and make choices using buttons, rather than directly
in the console as was done in the first iteration.